WIP: lang: moved EvalContext as Context into sym
