+++ builtin/tests.rs
@@ -9,7 +9,7 @@ use microcad_lang::{eval::*, sym::*};
 fn context_namespace() {
     use microcad_lang::src_ref::*;
 
-    let mut context = EvalContext::default();
+    let mut context = Context::default();
 
     let module = NamespaceBuilder::new("math")
         .add(Symbol::Value(
