+++ lang/eval/builtin_function.rs
@@ -3,10 +3,10 @@
 
 //! Builtin function evaluation entity
 
-use crate::{eval::*, parse::*, src_ref::SrcRef};
+use crate::{eval::*, parse::*, src_ref::SrcRef, sym::*};
 
 /// Type of the functor which receives a call
-pub type BuiltinFunctionFn = dyn Fn(&ArgumentMap, &mut EvalContext) -> EvalResult<Option<Value>>;
+pub type BuiltinFunctionFn = dyn Fn(&ArgumentMap, &mut Context) -> EvalResult<Option<Value>>;
 
 /// Builtin function
 #[derive(Clone)]
@@ -43,11 +43,7 @@ impl CallTrait for BuiltinFunction {
     /// # Arguments
     /// - `args`: Function arguments
     /// - `context`: Execution context
-    fn call(
-        &self,
-        args: &CallArgumentList,
-        context: &mut EvalContext,
-    ) -> EvalResult<Option<Value>> {
+    fn call(&self, args: &CallArgumentList, context: &mut Context) -> EvalResult<Option<Value>> {
         let arg_map = args.get_matching_arguments(context, &self.signature.parameters)?;
         let result = (self.f)(&arg_map, context)?;
