+++ lang/parse/assignment.rs
@@ -20,7 +20,7 @@ pub struct Assignment {
 
 impl Assignment {
     /// Make a symbol from the assignment
-    pub fn make_symbol(&self, context: &mut EvalContext) -> EvalResult<Symbol> {
+    pub fn make_symbol(&self, context: &mut Context) -> EvalResult<Symbol> {
         Ok(Symbol::Value(
             self.name.id().clone(),
             self.value.eval(context)?,
@@ -69,7 +69,7 @@ impl Parse for Assignment {
 impl Eval for Assignment {
     type Output = Symbol;
 
-    fn eval(&self, context: &mut EvalContext) -> EvalResult<Self::Output> {
+    fn eval(&self, context: &mut Context) -> EvalResult<Self::Output> {
         let symbol = self.make_symbol(context)?;
         context.add(symbol.clone());
         Ok(symbol)
