+++ lang/parse/call/method_call.rs
@@ -3,7 +3,7 @@
 
 //! Method call
 
-use crate::{eval::*, parse::*, parser::*, src_ref::*};
+use crate::{eval::*, parse::*, parser::*, src_ref::*, sym::*};
 
 /// Method call
 #[derive(Clone, Debug)]
@@ -18,7 +18,7 @@ pub struct MethodCall {
 
 impl MethodCall {
     /// Evaluate the method call in a context
-    pub fn eval(&self, context: &mut EvalContext, lhs: &Expression) -> EvalResult<Value> {
+    pub fn eval(&self, context: &mut Context, lhs: &Expression) -> EvalResult<Value> {
         use call::call_method::CallMethod;
 
         match lhs.eval(context)? {
