+++ lang/parse/expression/nested.rs
@@ -3,7 +3,7 @@
 
 //! Nested item list parser entity
 
-use crate::{eval::*, parse::*, parser::*, src_ref::*};
+use crate::{eval::*, parse::*, parser::*, src_ref::*, sym::*};
 
 /// Nested item list, e.g. an expression like `foo bar() {}`
 #[derive(Clone, Debug)]
@@ -47,7 +47,7 @@ impl SrcReferrer for Nested {
 impl Eval for Nested {
     type Output = Option<Value>;
 
-    fn eval(&self, context: &mut EvalContext) -> EvalResult<Self::Output> {
+    fn eval(&self, context: &mut Context) -> EvalResult<Self::Output> {
         let mut nodes = Vec::new();
 
         for item in self.0.iter() {
