+++ lang/parse/format_string/mod.rs
@@ -9,7 +9,7 @@ mod format_spec;
 pub use format_expression::*;
 pub use format_spec::*;
 
-use crate::{eval::*, parse::*, parser::*, src_ref::*};
+use crate::{eval::*, parse::*, parser::*, src_ref::*, sym::*};
 
 /// Format string item
 #[derive(Clone, Debug)]
@@ -79,7 +79,7 @@ impl std::fmt::Display for FormatString {
 impl Eval for FormatString {
     type Output = Value;
 
-    fn eval(&self, context: &mut EvalContext) -> EvalResult<Value> {
+    fn eval(&self, context: &mut Context) -> EvalResult<Value> {
         let mut result = String::new();
         for elem in &self.0 {
             match elem {
@@ -123,7 +123,7 @@ fn simple_string() {
 
     let s = FormatString::parse(pair).expect("test error");
     assert_eq!(s.section_count(), 1);
-    let mut context = EvalContext::default();
+    let mut context = Context::default();
     let value = s.eval(&mut context).expect("test error");
 
     assert_eq!(value, Value::String(Refer::none("Hello, World!".into())));
@@ -142,7 +142,7 @@ fn format_string() {
 
     let s = FormatString::parse(pair).expect("test error");
     assert_eq!(s.section_count(), 3);
-    let mut context = EvalContext::default();
+    let mut context = Context::default();
     let value = s.eval(&mut context).expect("test error");
 
     assert_eq!(value, Value::String(Refer::none("A6B".into())));
