+++ lang/parse/function/function_statement.rs
@@ -3,7 +3,7 @@
 
 //! Function statement parser entity
 
-use crate::{eval::*, parse::*, parser::*, src_ref::*};
+use crate::{eval::*, parse::*, parser::*, src_ref::*, sym::*};
 
 /// Function statement
 #[derive(Clone, Debug)]
@@ -92,7 +92,7 @@ impl Parse for FunctionStatement {
 impl Eval for FunctionStatement {
     type Output = Option<Value>;
 
-    fn eval(&self, context: &mut EvalContext) -> EvalResult<Self::Output> {
+    fn eval(&self, context: &mut Context) -> EvalResult<Self::Output> {
         match self {
             FunctionStatement::Assignment(assignment) => {
                 assignment.eval(context)?;
