+++ lang/parse/module/node_body.rs
@@ -40,7 +40,7 @@ impl Parse for NodeMarker {
 impl Eval for NodeMarker {
     type Output = Option<ObjectNode>;
 
-    fn eval(&self, context: &mut EvalContext) -> EvalResult<Self::Output> {
+    fn eval(&self, context: &mut Context) -> EvalResult<Self::Output> {
         match self.name.to_string().as_str() {
             "children" => Ok(Some(crate::objects::ObjectNode::new(
                 crate::objects::ObjectNodeInner::ChildrenNodeMarker,
@@ -104,7 +104,7 @@ impl Parse for NodeBodyStatement {
 impl Eval for NodeBodyStatement {
     type Output = Option<Value>;
 
-    fn eval(&self, context: &mut EvalContext) -> EvalResult<Self::Output> {
+    fn eval(&self, context: &mut Context) -> EvalResult<Self::Output> {
         match self {
             Self::NodeMarker(marker) => Ok(marker.eval(context)?.map(Value::Node)),
             Self::Use(use_statement) => {
@@ -178,7 +178,7 @@ impl Parse for NodeBody {
 impl Eval for NodeBody {
     type Output = crate::objects::ObjectNode;
 
-    fn eval(&self, context: &mut EvalContext) -> EvalResult<Self::Output> {
+    fn eval(&self, context: &mut Context) -> EvalResult<Self::Output> {
         let mut group = crate::objects::group();
 
         for statement in &self.statements {
