+++ lang/parse/namespace/namespace_statement.rs
@@ -60,7 +60,7 @@ impl Parse for NamespaceStatement {
 impl Eval for NamespaceStatement {
     type Output = ();
 
-    fn eval(&self, context: &mut EvalContext) -> std::result::Result<Self::Output, EvalError> {
+    fn eval(&self, context: &mut Context) -> std::result::Result<Self::Output, EvalError> {
         match self {
             Self::Use(use_statement) => {
                 if let Some(symbols) = use_statement.eval(context)? {
